response status code  and env file

+++ src/controllers/blogsComment.ts
@@ -3,10 +3,9 @@ import blogComment from '../models/blogComment'
 
 
 const createComment=async(req:Request,res:Response)=>{
-    /* console.log(req.body.user)*/
+
     try{
         let user=req.user
-        console.log(req.user)
         if(user){
             
             const newComment=new blogComment({
@@ -16,42 +15,53 @@ const createComment=async(req:Request,res:Response)=>{
                 })
         
                 const comment= await newComment.save()
-                res.status(200).json({message:"Your comment to this blog is successful sent"})
+                return res.status(200).json({message:"Your comment to this blog is successful sent"})
         }else{
-            res.status(400).json({error:"we are validating you before you place a comment and something went wrong"})   
+           return res.status(401).json({error:"we are validating you before you place a comment and something went wrong"})   
         }
        
    
     }catch{
-        res.status(500).json({error:"some thing went wrong with your comment"})
+       return res.status(500).json({error:"some thing went wrong with your comment"})
     }
 }
 
 const readComment=async(req:Request,res:Response)=>{
     try {
         const post = await blogComment.find({blogId: req.params.id })
-        res.status(200).json({data:post})
+        return res.status(200).json({data:post})
     } catch {
-        res.status(400).json({ error: "comments doesn't exist!" })
+       return res.status(404).json({ error: "comments doesn't exist!" })
     }
 }
 
 const singleComment=async(req:Request,res:Response)=>{
     try {
         const comment=await blogComment.findOne({ _id: req.params.comment_id })
-        res.status(200).json({data:comment})
+        if(comment){
+            return res.status(200).json({data:comment})
+        }else{
+            return res.status(404).json({error:"Comment doesn't exist!"})
+        }
+        
     } catch {
-        res.status(400).json({ error: "Comment doesn't exist!" })
+        return res.status(404).json({ error: "Comment doesn't exist!" })
     } 
 }
 
 
 const deleteComment=async(req:Request,res:Response)=>{
     try {
-        await blogComment.deleteOne({ _id: req.params.comment_id })
-        res.status(204).send()
+        const comment=await blogComment.findOne({ _id: req.params.comment_id })
+        if(comment){
+            await blogComment.deleteOne({ _id: req.params.comment_id })
+           return res.status(204).json({message:"comment is deleted"}) 
+        }else{
+           return res.status(400).json({error:"comment doesn't exist"}) 
+        }
+      
     } catch {
-        res.status(400).json({ error: "Comment doesn't exist!" })
+       return  res.status(404).json({ error: "Comment doesn't exist!" })
     }
 }
