+++ src/controllers/querriesController.ts
@@ -11,30 +11,36 @@ const createQuerries=async(req:Request,res:Response)=>{
             client_message:req.body.message 
         })
         const query=await clientQuery.save();
-        res.status(200).json({message:"Querries is successfully created"})
+        return res.status(200).json({message:"Querries is successfully created"})
 
     } catch (error) {
-        res.status(500).json({error:"internal server error"})
+       return res.status(500).json({error:"internal server error"})
     }
 }
 
 const getQuerries=async(req:Request,res:Response)=>{
     const allqueries=await Clients.find();
-    if(allqueries){
-        res.status(200).json({data:allqueries})
+    if(allqueries.length>0){
+        return res.status(200).json({data:allqueries})
     }else{
-        res.status(400).json({"message":"no client querries found in record"}) 
+       return res.status(404).json({error:"no client querries found in record"}) 
     }
 
 }
 
 const deleteQuerries=async(req:Request,res:Response)=>{
     try {
-        await Clients.deleteOne({ _id: req.params.id })
-        res.status(204).json({message:"Client Querry is deleted"})
-       // res.status(204)
+        const querry=await Clients.findOne({_id: req.params.id})
+        if(querry){
+            await Clients.deleteOne({ _id: req.params.id })
+           return res.status(204).json({message:"Client Querry is deleted"})
+        }else{
+            return res.status(404).json({error:"Client Query is not exist"})
+        }
+        
+      
     } catch {
-        res.status(400).json({ error: "Querry doesn't exist!" })
+        return res.status(404).json({ error: "Querry doesn't exist!" })
     }
 }
